Introduce separate test mixin per adapter method. Fix bug when using compacters with `AdapterSetup` context (now covered by tests).
