+++ tests_adapters/test_bert.py
@@ -4,16 +4,14 @@ from tests.bert.test_modeling_bert import *
 from transformers import BertAdapterModel
 from transformers.testing_utils import require_torch
 
+from .methods import BottleneckAdapterTestMixin, CompacterTestMixin, PrefixTuningTestMixin
 from .test_adapter import AdapterTestBase, make_config
 from .test_adapter_backward_compability import CompabilityTestMixin
-from .test_adapter_common import AdapterModelTestMixin
-from .test_adapter_compacter import CompacterTestMixin
 from .test_adapter_composition import ParallelAdapterInferenceTestMixin, ParallelTrainingMixin
 from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_embeddings import EmbeddingTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
-from .test_adapter_training import AdapterTrainingTestMixin
 from .test_common import AdapterModelTesterMixin
 
 
@@ -38,13 +36,13 @@ class BertAdapterTestBase(AdapterTestBase):
 
 @require_torch
 class BertAdapterTest(
-    EmbeddingTestMixin,
+    BottleneckAdapterTestMixin,
     CompacterTestMixin,
-    AdapterModelTestMixin,
+    PrefixTuningTestMixin,
+    EmbeddingTestMixin,
     AdapterFusionModelTestMixin,
     CompabilityTestMixin,
     PredictionHeadModelTestMixin,
-    AdapterTrainingTestMixin,
     ParallelAdapterInferenceTestMixin,
     ParallelTrainingMixin,
     BertAdapterTestBase,
