+++ tests_adapters/test_t5.py
@@ -6,16 +6,14 @@ from tests.t5.test_modeling_t5 import *
 from transformers import T5AdapterModel
 from transformers.testing_utils import require_torch
 
+from .methods import BottleneckAdapterTestMixin, CompacterTestMixin, PrefixTuningTestMixin
 from .test_adapter import AdapterTestBase, make_config
 from .test_adapter_backward_compability import CompabilityTestMixin
-from .test_adapter_common import AdapterModelTestMixin
-from .test_adapter_compacter import CompacterTestMixin
 from .test_adapter_composition import ParallelAdapterInferenceTestMixin, ParallelTrainingMixin
 from .test_adapter_conversion import ModelClassConversionTestMixin
 from .test_adapter_embeddings import EmbeddingTestMixin
 from .test_adapter_fusion_common import AdapterFusionModelTestMixin
 from .test_adapter_heads import PredictionHeadModelTestMixin
-from .test_adapter_training import AdapterTrainingTestMixin
 from .test_common import AdapterModelTesterMixin
 
 
@@ -81,17 +79,16 @@ class T5AdapterTestBase(AdapterTestBase):
 
 @require_torch
 class T5AdapterTest(
-    T5AdapterTestBase,
+    BottleneckAdapterTestMixin,
     CompacterTestMixin,
+    PrefixTuningTestMixin,
     EmbeddingTestMixin,
     CompabilityTestMixin,
     ParallelAdapterInferenceTestMixin,
     ParallelTrainingMixin,
-    AdapterModelTestMixin,
     AdapterFusionModelTestMixin,
-    AdapterTrainingTestMixin,
     PredictionHeadModelTestMixin,
-    AdapterTestBase,
+    T5AdapterTestBase,
     unittest.TestCase,
 ):
     pass
